rename dt-list-help-box to dt-list-filters and added splitBy variant of it.
