+++ src/components/layout/index.js
@@ -1,5 +1,5 @@
 export { DtAlert } from './dt-alert/dt-alert.js';
 export { DtList } from './dt-list/dt-list.js';
-export { DtListHelpBox } from './dt-list-help-box/dt-list-help-box.js';
+export { DtListHelpBox } from './dt-list-filters/dt-list-filters.js';
 export { DtModal } from './dt-modal/dt-modal.js';
 export { DtTile } from './dt-tile/dt-tile.js';
